feat(platform): canvas file explorer tree by source directory
Replace the canvas tab strip + bottom output dock with a directory
tree in the Workspace-files pane's visual language: code/ (AI files),
uploads/ (user attachments), output/ (code-run outputs), each a
collapsible folder with per-extension file icons. Provenance is now
the file's place in the tree instead of a per-tab icon nobody could
read. The shared row primitive, icon map, and tree keyboard
navigation move to file-tree-primitives.tsx so both panes render the
same tree.
