+++ services/platform/messages/en.json
@@ -1831,11 +1831,8 @@
       "fileActionsAriaLabel": "File actions",
       "fileSizeAriaLabel": "File size",
       "sourceRunOutput": "Code output",
-      "outputExpand": "Show all files",
-      "outputCollapse": "Show fewer files",
       "sourceAgentWrite": "AI files",
       "sourceUserUpload": "Uploaded",
-      "noFilesYet": "No files yet",
       "writing": "Writing…",
       "runStatus": {
         "queued": "Queued",
